+++ src/main/java/grupo2/SistemaAutomotor/servicio/boleta/IBoletaServicio.java
@@ -4,6 +4,7 @@ import grupo2.SistemaAutomotor.modelo.Automotor;
 import grupo2.SistemaAutomotor.modelo.Boleta;
 import grupo2.SistemaAutomotor.modelo.Municipio;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface IBoletaServicio {
@@ -12,7 +13,7 @@ public interface IBoletaServicio {
     void guardarBoleta(Boleta boleta);
     void eliminarBoleta(Integer id);
     List<Boleta> buscarBoletasPorDominio(Automotor automotor);
-    float sumarBoletasPorMunicipio(Municipio municipio);
+    BigDecimal sumarBoletasPorMunicipio(Municipio municipio);
     List<Boleta> buscarBoletasPorMunicipio(Municipio municipio);
     List<Boleta> buscarBoletasPorDomonioYFechaDesde(Automotor automotor, boolean estado, int fecha);
